backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==1.4.49
pydantic==1.10.9
alembic==1.11.1
openai>=1.3.0
python-dotenv==1.0.0